requirements.txt
pyquery
pyyaml
requests
retry 
